Add questions and answers

+++ src/Controller/QuestionController.php
@@ -2,39 +2,45 @@
 
 namespace App\Controller;
 
+use App\Entity\Answer;
 use App\Entity\Question;
+use App\Form\AnswerType;
 use App\Form\QuestionType;
+use App\Repository\AnswerRepository;
 use App\Repository\QuestionRepository;
+use DateTime;
+use Doctrine\ORM\EntityManagerInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 #[Route('/question')]
+#[IsGranted('ROLE_USER')]
 class QuestionController extends AbstractController
 {
     #[Route('/', name: 'question_index', methods: ['GET'])]
     public function index(
         QuestionRepository $questionRepository
     ): Response {
-        return $this->render(
-            'question/index.html.twig',
-            [
-                'questions' => $questionRepository->findAll(),
-            ]
-        );
+        $questions = $questionRepository->findAll();
+
+        return $this->render('question/index.html.twig', ['questions' => $questions]);
     }
 
     #[Route('/new', name: 'question_new', methods: ['GET', 'POST'])]
     public function new(
-        Request $request
+        Request $request,
+        EntityManagerInterface $entityManager
     ): Response {
         $question = new Question();
         $form = $this->createForm(QuestionType::class, $question);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
+            $question->setUser($this->getUser())
+                     ->setDateTime(new DateTime());
             $entityManager->persist($question);
             $entityManager->flush();
 
@@ -50,14 +56,33 @@ class QuestionController extends AbstractController
         );
     }
 
-    #[Route('/{id}', name: 'question_show', methods: ['GET'])]
+    #[Route('/{id}', name: 'question_show', methods: ['GET', 'POST'])]
     public function show(
-        Question $question
+        Request $request,
+        Question $question,
+        AnswerRepository $answerRepository,
+        EntityManagerInterface $entityManager
     ): Response {
+        $answer = new Answer();
+        $answer->setUser($this->getUser());
+        $form = $this->createForm(AnswerType::class, $answer);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $question->addAnswer($answer->setDateTime(new DateTime()));
+            $entityManager->flush();
+            $entityManager->refresh($question);
+        }
+
+        $answers = $answerRepository->getLatest($question);
+        $entityManager->clear(Answer::class);
+
         return $this->render(
             'question/show.html.twig',
             [
                 'question' => $question,
+                'answers'  => $answers,
+                'form'     => $form->createView(),
             ]
         );
     }
@@ -65,13 +90,20 @@ class QuestionController extends AbstractController
     #[Route('/{id}/edit', name: 'question_edit', methods: ['GET', 'POST'])]
     public function edit(
         Request $request,
-        Question $question
+        Question $question,
+        EntityManagerInterface $entityManager
     ): Response {
+        if (!$this->isGranted('ROLE_SUPERADMIN') && $this->getUser()->getUsername() !== $question->getUser()->getUsername()) {
+            $this->addFlash('user_error', 'Nije dozvoljena izmjena upita.');
+
+            return $this->redirectToRoute('question_index');
+        }
+
         $form = $this->createForm(QuestionType::class, $question);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            $entityManager->flush();
 
             return $this->redirectToRoute('question_index');
         }
@@ -88,10 +120,16 @@ class QuestionController extends AbstractController
     #[Route('/{id}', name: 'question_delete', methods: ['DELETE'])]
     public function delete(
         Request $request,
-        Question $question
+        Question $question,
+        EntityManagerInterface $entityManager
     ): Response {
+        if (!$this->isGranted('ROLE_SUPERADMIN') && $this->getUser()->getUsername() !== $question->getUser()->getUsername()) {
+            $this->addFlash('user_error', 'Nije dozvoljeno brisanje upita.');
+
+            return $this->redirectToRoute('question_index');
+        }
+
         if ($this->isCsrfTokenValid('delete' . $question->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($question);
             $entityManager->flush();
         }
